+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/fargate/_01_merge.py
@@ -1,16 +1,23 @@
 import polars as pl
 
 import projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate.utils.merge_utils as mUtils
-from polars_utils import expressions as expr
 from polars_utils import utils
 from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
     AscwdsWorkplaceCleanedColumns as AWPClean,
+)
+from utils.column_names.cleaned_data_files.ascwds_workplace_job_roles import (
+    AscwdsWorkplaceJobRolesColumns as AWPJobRoles,
 )
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 
 workplace_columns = [
     AWPClean.establishment_id,
     AWPClean.ascwds_workplace_import_date,
+    AWPJobRoles.job_role_code,
+    AWPJobRoles.employees,
+    AWPJobRoles.starters,
+    AWPJobRoles.leavers,
+    AWPJobRoles.vacancies,
 ]
 
 metadata_columns = [
@@ -69,10 +76,7 @@ def main(
         source=job_role_estimates_source, selected_columns=job_role_estimates_columns
     )
     cleaned_ascwds_workplace_lf = utils.scan_parquet(
-        prepared_slv_dataset_source
-    ).select(
-        *[AWPClean.establishment_id, AWPClean.ascwds_workplace_import_date],
-        expr.is_slv_job_role_column()
+        prepared_slv_dataset_source, selected_columns=workplace_columns
     )
 
     # TODO: Placeholder only