+++ utils/column_names/cleaned_data_files/ascwds_workplace_job_roles.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class AscwdsWorkplaceJobRolesColumns:
+    job_role_code: str = "job_role_code"
+    employees: str = "employees"
+    starters: str = "starters"
+    leavers: str = "leavers"
+    vacancies: str = "vacancies"